split evm-only loadtest harness files
